Move SchemaRegistry into its own namespace
- Split SchemaRegistry into multiple definitions (is it a good idea?)
