+++ src/AvroSourceGenerator/Parsing/Parser.cs
@@ -2,8 +2,8 @@
 using System.Runtime.CompilerServices;
 using System.Text.Json;
 using AvroSourceGenerator.Diagnostics;
+using AvroSourceGenerator.Registry.Extensions;
 using AvroSourceGenerator.Schemas;
-using AvroSourceGenerator.Schemas.Extensions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
